+++ src/se/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * This package contains the classes for graph.
+ * <p>
+ * The main components of this package include:
+ * </p>
+ * <ul>
+ *   <li>{@link se.lab.graph.Edge} - Represents an edge in the graph.</li>
+ *   <li>{@link se.lab.graph.Vertex} - Represents a vertex in the graph.</li>
+ *   <li>{@link se.lab.graph.MyGraph} - Represents graph.</li>
+ * </ul>
+ * For more details, refer to the individual class documentation.
+ *
+ * @since 1.0
+ */
+
+package se;
